naming consistency across optimizers
